+++ example/lib/main.dart
@@ -38,7 +38,6 @@ class ExampleArcaneApp extends StatelessWidget {
             home: HomeScreen(),
             showPerformanceOverlay: false,
             theme: ArcaneTheme(
-                blurMode: ArcaneBlurMode.frost,
                 themeMode: ThemeMode.system,
                 scheme: ContrastedColorScheme.fromScheme(ColorSchemes.violet)),
           ));