+++ example/lib/screen/home.dart
@@ -11,49 +11,98 @@ class HomeScreen extends StatefulWidget {
 
 class _HomeScreenState extends State<HomeScreen> {
   @override
-  Widget build(BuildContext context) => ArcaneScreen(
-        fab: Fab(
-            child: const Icon(Icons.plus),
-            onPressed: () {
-              setState(() {
-                list.add(
-                    "Item ${list.length} ${DateTime.now().toIso8601String()}");
-              });
-            }),
-        header: Bar(
-          titleText: "Thing",
-          trailing: [],
-        ),
-        child: Collection(
-          children: [
-            Section(
-                headerText: "Section 1",
-                child: SGridView.builder(
-                  builder: (context, i) => ContextMenu(
-                      child: Card(
-                        child: Text("Item $i"),
-                        onPressed: () => DialogConfirm(
-                          title: "Confirm",
-                          onConfirm: () {},
-                        ).open(context),
-                        fillColor: [
-                          Colors.blue,
-                          Colors.red,
-                          Colors.green
-                        ][i % 3],
-                        filled: true,
-                      ),
-                      items: [
-                        MenuButton(child: Text("A")),
-                        MenuButton(child: Text("b")),
-                        MenuButton(child: Text("c")),
-                        MenuButton(child: Text("d")),
-                        MenuButton(child: Text("e")),
-                        MenuButton(child: Text("f")),
-                      ]),
-                  childCount: 10,
-                ))
-          ],
+  Widget build(BuildContext context) => MutablePylon<int>(
+        rebuildChildren: true,
+        local: true,
+        value: 0,
+        builder: (context) => NavigationScreen(
+          type: NavigationType.bottomNavigationBar,
+          index: context.pylon<int>(),
+          onIndexChanged: (index) => context.setPylon<int>(index),
+          tabs: <NavItem>[
+            NavTab(
+                label: "Sliver Screen",
+                icon: Icons.airplane,
+                selectedIcon: Icons.airplane_fill,
+                builder: (context) => SliverScreen(
+                    gutter: false,
+                    header: const Bar(
+                      titleText: "This is bar 1",
+                    ),
+                    sliver: SListView.builder(
+                      childCount: 1000,
+                      builder: (context, i) => BasicCard(
+                        title: Text("Tile $i"),
+                      ).padBottom(8),
+                    ))),
+            NavTab(
+                label: "Sliver Screen",
+                icon: Icons.airplane,
+                selectedIcon: Icons.airplane_fill,
+                builder: (context) => SliverScreen(
+                    gutter: false,
+                    header: const Bar(
+                      titleText: "This is bar 1",
+                    ),
+                    sliver: SListView.builder(
+                      childCount: 1000,
+                      builder: (context, i) => BasicCard(
+                        title: Text("Tile $i"),
+                      ).padBottom(8),
+                    ))),
+            NavTab(
+                label: "Sliver Screen",
+                icon: Icons.airplane,
+                selectedIcon: Icons.airplane_fill,
+                builder: (context) => SliverScreen(
+                    gutter: false,
+                    header: const Bar(
+                      titleText: "This is bar 1",
+                    ),
+                    sliver: SListView.builder(
+                      childCount: 1000,
+                      builder: (context, i) => BasicCard(
+                        title: Text("Tile $i"),
+                      ).padBottom(8),
+                    ))),
+            NavTab(
+                label: "Fill Screen",
+                icon: Icons.activity,
+                selectedIcon: Icons.activity_fill,
+                builder: (context) => const FillScreen(
+                    gutter: false,
+                    header: Bar(
+                      titleText: "This is bar 2",
+                    ),
+                    // using fill screen
+                    child: CenterBody(
+                        icon: Icons.activity, message: "This is Fill Screen"))),
+            NavTab(
+                label: "Sliver Fill",
+                icon: Icons.activity,
+                selectedIcon: Icons.activity_fill,
+                builder: (context) => const SliverScreen(
+                    gutter: false,
+                    header: Bar(
+                      titleText: "This is bar 5",
+                    ),
+                    // using fill screen
+                    sliver: SliverFill(
+                      child: CenterBody(
+                          icon: Icons.activity,
+                          message: "This is a Sliver Filled"),
+                    ))),
+          ].toList(),
+          sidebarHeader: (context) => const ArcaneSidebarHeader(
+            titleText: "Header",
+            subtitleText: "Subtitle",
+            trailing: [
+              Icon(Icons.search_ionic),
+            ],
+          ),
+          sidebarFooter: (context) => ArcaneSidebarFooter(
+            content: const Text("Footer"),
+          ),
         ),
       );
 }